Oit: integrated volume rasterizer and renderer
