+++ modules/oit/src/oitmodule.cpp
@@ -29,18 +29,20 @@
 
 #include <modules/oit/oitmodule.h>
 
-#include <inviwo/core/common/inviwomodule.h>               // for InviwoModule, Mod...
-#include <inviwo/core/ports/outportiterable.h>             // for OutportIterableIm...
-#include <inviwo/core/util/exception.h>                    // for Exception
-#include <modules/oit/ports/rasterizationport.h>           // for RasterizationOutport
-#include <modules/oit/processors/calcnormalsprocessor.h>   // for CalcNormalsProcessor
-#include <modules/oit/processors/linerasterizer.h>         // for LineRasterizer
-#include <modules/oit/processors/meshrasterizer.h>         // for MeshRasterizer
+#include <inviwo/core/common/inviwomodule.h>              // for InviwoModule, Mod...
+#include <inviwo/core/ports/outportiterable.h>            // for OutportIterableIm...
+#include <inviwo/core/util/exception.h>                   // for Exception
+#include <modules/oit/ports/rasterizationport.h>          // for RasterizationOutport
+#include <modules/oit/processors/calcnormalsprocessor.h>  // for CalcNormalsProcessor
+#include <modules/oit/processors/linerasterizer.h>        // for LineRasterizer
+#include <modules/oit/processors/meshrasterizer.h>        // for MeshRasterizer
+#include <modules/oit/processors/meshvolumerenderer.h>
 #include <modules/oit/processors/rasterizationrenderer.h>  // for RasterizationRend...
 #include <modules/oit/processors/rasterizer.h>
 #include <modules/oit/processors/sphererasterizer.h>        // for SphereRasterizer
 #include <modules/oit/processors/transformrasterization.h>  // for TransformRasteriz...
-#include <modules/opengl/shader/shadermanager.h>            // for ShaderManager
+#include <modules/oit/processors/volumerasterizer.h>
+#include <modules/opengl/shader/shadermanager.h>  // for ShaderManager
 
 #include <functional>  // for __base
 #include <memory>      // for unique_ptr
@@ -59,9 +61,11 @@ OITModule::OITModule(InviwoApplication* app) : InviwoModule(app, "Oit") {
     registerProcessor<LineRasterizer>();
     registerProcessor<MeshRasterizer>();
     registerProcessor<RasterizationRenderer>();
+    registerProcessor<MeshVolumeRenderer>();
     registerProcessor<TransformRasterization>();
     registerProcessor<CalcNormalsProcessor>();
     registerProcessor<SphereRasterizer>();
+    registerProcessor<VolumeRasterizer>();
 
     // Ports
     registerPort<RasterizationInport>();
